fix(patina): align no-unused-components registration parity
